Extracted page/rest/sit/stand/door commands from act_movement.cpp
