+++ src/act_movement.cpp
@@ -29,7 +29,6 @@
 #include "engine/db/global_objects.h"
 #include "gameplay/mechanics/liquid.h"
 #include "engine/core/utils_char_obj.inl"
-#include "gameplay/mechanics/treasure_cases.h"
 #include "gameplay/mechanics/sight.h"
 #include "gameplay/mechanics/awake.h"
 
@@ -1001,503 +1000,6 @@ void do_hidemove(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
 	if (!sneaking || IsAffectedBySpell(ch, ESpell::kGlitterDust)) {
 		RemoveAffectFromChar(ch, ESpell::kSneak);
 		AFF_FLAGS(ch).unset(EAffect::kSneak);
-	}
-}
-
-#define DOOR_IS_OPENABLE(ch, obj, door)    ((obj) ? \
-            ((GET_OBJ_TYPE(obj) == EObjType::kContainer) && \
-            OBJVAL_FLAGGED(obj, EContainerFlag::kCloseable)) :\
-            (EXIT_FLAGGED(EXIT(ch, door), EExitFlag::kHasDoor)))
-#define DOOR_IS(ch, door)    ((EXIT_FLAGGED(EXIT(ch, door), EExitFlag::kHasDoor)))
-
-#define DOOR_IS_OPEN(ch, obj, door)    ((obj) ? \
-            (!OBJVAL_FLAGGED(obj, EContainerFlag::kShutted)) :\
-            (!EXIT_FLAGGED(EXIT(ch, door), EExitFlag::kClosed)))
-#define DOOR_IS_BROKEN(ch, obj, door)    ((obj) ? \
-    (OBJVAL_FLAGGED(obj, EContainerFlag::kLockIsBroken)) :\
-    (EXIT_FLAGGED(EXIT(ch, door), EExitFlag::kBrokenLock)))
-#define DOOR_IS_UNLOCKED(ch, obj, door)    ((obj) ? \
-            (!OBJVAL_FLAGGED(obj, EContainerFlag::kLockedUp)) :\
-            (!EXIT_FLAGGED(EXIT(ch, door), EExitFlag::kLocked)))
-#define DOOR_IS_PICKPROOF(ch, obj, door) ((obj) ? \
-    (OBJVAL_FLAGGED(obj, EContainerFlag::kUncrackable) || OBJVAL_FLAGGED(obj, EContainerFlag::kLockIsBroken)) : \
-    (EXIT_FLAGGED(EXIT(ch, door), EExitFlag::kPickroof) || EXIT_FLAGGED(EXIT(ch, door), EExitFlag::kBrokenLock)))
-
-#define DOOR_IS_CLOSED(ch, obj, door)    (!(DOOR_IS_OPEN(ch, obj, door)))
-#define DOOR_IS_LOCKED(ch, obj, door)    (!(DOOR_IS_UNLOCKED(ch, obj, door)))
-#define DOOR_KEY(ch, obj, door)        ((obj) ? (GET_OBJ_VAL(obj, 2)) : \
-                    (EXIT(ch, door)->key))
-#define DOOR_LOCK_COMPLEX(ch, obj, door) ((obj) ? \
-            (GET_OBJ_VAL(obj,3)) :\
-            (EXIT(ch, door)->lock_complexity))
-
-int find_door(CharData *ch, const char *type, char *dir, EDoorScmd scmd) {
-	int door;
-	bool found = false;
-
-	if (*dir) {
-		if ((door = search_block(dir, dirs, false)) == -1
-			&& (door = search_block(dir, dirs_rus, false)) == -1) {
-			return kWrongDir;
-		}
-		if (EXIT(ch, door)) {
-			if (EXIT(ch, door)->keyword && EXIT(ch, door)->vkeyword) {
-				if (isname(type, EXIT(ch, door)->keyword) || isname(type, EXIT(ch, door)->vkeyword)) {
-					return (door);
-				} else {
-					return kWrongDirDoorName;
-				}
-			} else if (utils::IsAbbr(type, "дверь") || utils::IsAbbr(type, "door")) {
-				return door;
-			} else {
-				return kWrongDoorName;
-			}
-		} else {
-			return kNoDoorGivenDir;
-		}
-	} else {
-		if (!*type) {
-			return (kDoorNameIsEmpty);
-		}
-		for (door = 0; door < EDirection::kMaxDirNum; door++) {
-			found = false;
-			if (EXIT(ch, door)) {
-				if (EXIT(ch, door)->keyword && EXIT(ch, door)->vkeyword) {
-					if (isname(type, EXIT(ch, door)->keyword) || isname(type, EXIT(ch, door)->vkeyword)) {
-						found = true;
-					}
-				} else if (DOOR_IS(ch, door) && (utils::IsAbbr(type, "дверь") || utils::IsAbbr(type, "door"))) {
-					found = true;
-				}
-			}
-			if (found) {
-				if ((scmd == kScmdOpen && DOOR_IS_OPEN(ch, nullptr, door)) ||
-					(scmd == kScmdClose && DOOR_IS_CLOSED(ch, nullptr, door))) {
-					continue;
-				} else {
-					return door;
-				}
-			}
-		}
-		return (kWrongDoorName); //НЕПРАВИЛЬНО НАЗВАЛИ ДВЕРЬ БЕЗ УКАЗАНИЯ НАПРАВЛЕНИЯ
-	}
-
-}
-
-int HasKey(CharData *ch, ObjVnum key) {
-	for (auto o = ch->carrying; o; o = o->get_next_content()) {
-		if (GET_OBJ_VNUM(o) == key && key != -1) {
-			return true;
-		}
-	}
-
-	if (GET_EQ(ch, kHold)) {
-		if (GET_OBJ_VNUM(GET_EQ(ch, kHold)) == key && key != -1) {
-			return true;
-		}
-	}
-
-	return false;
-}
-
-const Bitvector kNeedOpen = 1 << 0;
-const Bitvector kNeedClosed = 1 << 1;
-const Bitvector kNeedUnlocked = 1 << 2;
-const Bitvector kNeedLocked = 1 << 3;
-
-const char *cmd_door[] =
-	{
-		"открыл$g",
-		"закрыл$g",
-		"отпер$q",
-		"запер$q",
-		"взломал$g"
-	};
-
-const char *a_cmd_door[] =
-	{
-		"открыть",
-		"закрыть",
-		"отпереть",
-		"запереть",
-		"взломать"
-	};
-
-const int flags_door[] =
-	{
-		kNeedClosed | kNeedUnlocked,
-		kNeedOpen,
-		kNeedClosed | kNeedLocked,
-		kNeedClosed | kNeedUnlocked,
-		kNeedClosed | kNeedLocked
-	};
-
-#define EXITN(room, door)        (world[room]->dir_option[door])
-
-inline void OPEN_DOOR(const RoomRnum room, ObjData *obj, const int door) {
-	if (obj) {
-		auto v = obj->get_val(1);
-		TOGGLE_BIT(v, EContainerFlag::kShutted);
-		obj->set_val(1, v);
-	} else {
-		TOGGLE_BIT(EXITN(room, door)->exit_info, EExitFlag::kClosed);
-	}
-}
-
-inline void LOCK_DOOR(const RoomRnum room, ObjData *obj, const int door) {
-	if (obj) {
-		auto v = obj->get_val(1);
-		TOGGLE_BIT(v, EContainerFlag::kLockedUp);
-		obj->set_val(1, v);
-	} else {
-		TOGGLE_BIT(EXITN(room, door)->exit_info, EExitFlag::kLocked);
-	}
-}
-
-void do_doorcmd(CharData *ch, ObjData *obj, int door, EDoorScmd scmd) {
-	bool deaf = false;
-	int other_room = 0;
-	int rev_dir[] = {EDirection::kSouth, EDirection::kWest, EDirection::kNorth, EDirection::kEast, EDirection::kDown, EDirection::kUp};
-	char local_buf[kMaxStringLength]; // строка, в которую накапливается совершенное действо
-	// пишем начало строки - кто чё сделал
-	sprintf(local_buf, "$n %s ", cmd_door[scmd]);
-
-	if (AFF_FLAGGED(ch, EAffect::kDeafness))
-		deaf = true;
-	// ищем парную дверь в другой клетке
-	RoomData::exit_data_ptr back;
-	if (!obj && EXIT(ch, door) && ((other_room = EXIT(ch, door)->to_room()) != kNowhere)) {
-		back = world[other_room]->dir_option[rev_dir[door]];
-		if (back) {
-			if ((back->to_room() != ch->in_room)
-				|| ((EXITDATA(ch->in_room, door)->exit_info
-					^ EXITDATA(other_room, rev_dir[door])->exit_info)
-					& (EExitFlag::kHasDoor | EExitFlag::kClosed | EExitFlag::kLocked))) {
-				back.reset();
-			}
-		}
-	}
-
-	switch (scmd) {
-		case kScmdOpen:
-		case kScmdClose:
-			if (scmd == kScmdOpen && obj && !open_otrigger(obj, ch, false))
-				return;
-			if (scmd == kScmdOpen && !obj && !open_wtrigger(world[ch->in_room], ch, door, false))
-				return;
-			if (scmd == kScmdOpen && !obj && back && !open_wtrigger(world[other_room], ch, rev_dir[door], false))
-				return;
-			if (scmd == kScmdClose && obj && !close_otrigger(obj, ch, false))
-				return;
-			if (scmd == kScmdClose && !obj && !close_wtrigger(world[ch->in_room], ch, door, false))
-				return;
-			if (scmd == kScmdClose && !obj && back && !close_wtrigger(world[other_room], ch, rev_dir[door], false))
-				return;
-			OPEN_DOOR(ch->in_room, obj, door);
-			if (back) {
-				OPEN_DOOR(other_room, obj, rev_dir[door]);
-			}
-			// вываливание и пурж кошелька
-			if (obj && system_obj::is_purse(obj)) {
-				sprintf(buf,
-						"<%s> {%d} открыл трупный кошелек %s.",
-						ch->get_name().c_str(),
-						GET_ROOM_VNUM(ch->in_room),
-						GetPlayerNameByUnique(GET_OBJ_VAL(obj, 3)));
-				mudlog(buf, NRM, kLvlGreatGod, MONEY_LOG, true);
-				system_obj::process_open_purse(ch, obj);
-				return;
-			} else {
-				SendMsgToChar(OK, ch);
-			}
-			break;
-
-		case kScmdUnlock:
-		case kScmdLock:
-			if (scmd == kScmdUnlock && obj && !open_otrigger(obj, ch, true))
-				return;
-			if (scmd == kScmdLock && obj && !close_otrigger(obj, ch, true))
-				return;
-			if (scmd == kScmdUnlock && !obj && !open_wtrigger(world[ch->in_room], ch, door, true))
-				return;
-			if (scmd == kScmdLock && !obj && !close_wtrigger(world[ch->in_room], ch, door, true))
-				return;
-			if (scmd == kScmdUnlock && !obj && back && !open_wtrigger(world[other_room], ch, rev_dir[door], true))
-				return;
-			if (scmd == kScmdLock && !obj && back && !close_wtrigger(world[other_room], ch, rev_dir[door], true))
-				return;
-			LOCK_DOOR(ch->in_room, obj, door);
-			if (back)
-				LOCK_DOOR(other_room, obj, rev_dir[door]);
-			if (!deaf)
-				SendMsgToChar("*Щелк*\r\n", ch);
-			if (obj) {
-				treasure_cases::UnlockTreasureCase(ch, obj);
-			}
-			break;
-
-		case kScmdPick:
-			if (obj && !pick_otrigger(obj, ch))
-				return;
-			if (!obj && !pick_wtrigger(world[ch->in_room], ch, door))
-				return;
-			if (!obj && back && !pick_wtrigger(world[other_room], ch, rev_dir[door]))
-				return;
-			LOCK_DOOR(ch->in_room, obj, door);
-			if (back)
-				LOCK_DOOR(other_room, obj, rev_dir[door]);
-			SendMsgToChar("Замок очень скоро поддался под вашим натиском.\r\n", ch);
-			strcpy(local_buf, "$n умело взломал$g ");
-			break;
-	}
-
-	// Notify the room
-	sprintf(local_buf + strlen(local_buf), "%s.", (obj) ? "$o3" : (EXIT(ch, door)->vkeyword ? "$F" : "дверь"));
-	if (!obj || (obj->get_in_room() != kNowhere)) {
-		act(local_buf, false, ch, obj, obj ? nullptr : EXIT(ch, door)->vkeyword, kToRoom);
-	}
-
-	// Notify the other room
-	if ((scmd == kScmdOpen || scmd == kScmdClose) && back) {
-		const auto &people = world[EXIT(ch, door)->to_room()]->people;
-		if (!people.empty()) {
-			sprintf(local_buf + strlen(local_buf) - 1, " с той стороны.");
-			int allowed_items_remained = 1000;
-			for (const auto to : people) {
-				if (0 == allowed_items_remained) {
-					break;
-				}
-				perform_act(local_buf, ch, obj, obj ? nullptr : EXIT(ch, door)->vkeyword, to);
-				--allowed_items_remained;
-			}
-		}
-	}
-}
-
-bool ok_pick(CharData *ch, ObjVnum /*keynum*/, ObjData *obj, int door, int scmd) {
-	const bool pickproof = DOOR_IS_PICKPROOF(ch, obj, door);
-
-	if (scmd != kScmdPick) {
-		return true;
-	}
-
-	if (pickproof) {
-		SendMsgToChar("Вы никогда не сможете взломать ЭТО.\r\n", ch);
-		return false;
-	}
-
-	if (!check_moves(ch, kPicklockMoves)) {
-		return false;
-	}
-
-	const PickProbabilityInformation &pbi = get_pick_probability(ch, DOOR_LOCK_COMPLEX(ch, obj, door));
-
-	const bool pick_success = pbi.unlock_probability >= number(1, 100);
-
-	if (pbi.skill_train_allowed) {
-		TrainSkill(ch, ESkill::kPickLock, pick_success, nullptr);
-	}
-
-	if (!pick_success) {
-		// неудачная попытка взлома - есть шанс сломать замок
-		const int brake_probe = number(1, 100);
-		if (pbi.brake_lock_probability >= brake_probe) {
-			SendMsgToChar("Вы все-таки сломали этот замок...\r\n", ch);
-			if (obj) {
-				auto v = obj->get_val(1);
-				SET_BIT(v, EContainerFlag::kLockIsBroken);
-				obj->set_val(1, v);
-			} else {
-				SET_BIT(EXIT(ch, door)->exit_info, EExitFlag::kBrokenLock);
-			}
-		} else {
-			if (pbi.unlock_probability == 0) {
-				SendMsgToChar("С таким сложным замком даже и пытаться не следует...\r\n", ch);
-				return false;
-			} else {
-				SendMsgToChar("Взломщик из вас пока еще никудышний.\r\n", ch);
-			}
-		}
-	}
-
-	return pick_success;
-}
-
-void do_gen_door(CharData *ch, char *argument, int, int subcmd) {
-	if (AFF_FLAGGED(ch, EAffect::kBlind)) {
-		SendMsgToChar("Очнитесь, вы же слепы!\r\n", ch);
-		return;
-	}
-
-	if (subcmd == kScmdPick && !ch->GetSkill(ESkill::kPickLock)) {
-		SendMsgToChar("Это умение вам недоступно.\r\n", ch);
-		return;
-	}
-	skip_spaces(&argument);
-	if (!*argument) {
-		sprintf(buf, "%s что?\r\n", a_cmd_door[subcmd]);
-		SendMsgToChar(CAP(buf), ch);
-		return;
-	}
-	char type[kMaxInputLength], dir[kMaxInputLength];
-	two_arguments(argument, type, dir);
-	int where_bits = EFind::kObjInventory | EFind::kObjRoom | EFind::kObjEquip;
-	if (isname(dir, "земля комната room ground"))
-		where_bits = EFind::kObjRoom;
-	else if (isname(dir, "инвентарь inventory"))
-		where_bits = EFind::kObjInventory;
-	else if (isname(dir, "экипировка equipment"))
-		where_bits = EFind::kObjEquip;
-
-	auto door = find_door(ch, type, dir, (EDoorScmd) subcmd);
-	ObjData *obj{nullptr};
-	CharData *victim{nullptr};
-	if (door < 0)
-		if (!generic_find(type, where_bits, ch, &victim, &obj)) {
-			switch (door) {
-				case kWrongDir:
-					SendMsgToChar("Уточните направление.\r\n", ch);
-					break;
-				case kWrongDirDoorName:
-					sprintf(buf, "Вы не видите '%s' в этой комнате.\r\n", type);
-					SendMsgToChar(buf, ch);
-					break;
-				case kNoDoorGivenDir:
-					sprintf(buf, "Вы не можете это '%s'.\r\n", a_cmd_door[subcmd]);
-					SendMsgToChar(buf, ch);
-					break;
-				case kDoorNameIsEmpty:
-					sprintf(buf, "Что вы хотите '%s'?\r\n", a_cmd_door[subcmd]);
-					SendMsgToChar(buf, ch);
-					break;
-				case kWrongDoorName:
-					sprintf(buf, "Вы не видите здесь '%s'.\r\n", type);
-					SendMsgToChar(buf, ch);
-					break;
-				default:
-					SendMsgToChar("Что-то с дверью произошло непнятное, сообщите богам.", ch);
-					break;
-			}
-			return;
-		}
-	if ((obj) || (door >= 0)) {
-		if ((obj) && !IS_IMMORTAL(ch) && (obj->has_flag(EObjFlag::kNamed))
-			&& NamedStuff::check_named(ch, obj, true))//Именной предмет открывать(закрывать) может только владелец
-		{
-			if (!NamedStuff::wear_msg(ch, obj))
-				SendMsgToChar("Просьба не трогать! Частная собственность!\r\n", ch);
-			return;
-		}
-		auto keynum = DOOR_KEY(ch, obj, door);
-		if ((subcmd == kScmdClose || subcmd == kScmdLock) && !ch->IsNpc() && NORENTABLE(ch))
-			SendMsgToChar("Ведите себя достойно во время боевых действий!\r\n", ch);
-		else if (!(DOOR_IS_OPENABLE(ch, obj, door)))
-			act("Вы никогда не сможете $F это!", false, ch, nullptr, a_cmd_door[subcmd], kToChar);
-		else if (!DOOR_IS_OPEN(ch, obj, door)
-			&& IS_SET(flags_door[subcmd], kNeedOpen))
-			SendMsgToChar("Вообще-то здесь закрыто!\r\n", ch);
-		else if (!DOOR_IS_CLOSED(ch, obj, door) && IS_SET(flags_door[subcmd], kNeedClosed))
-			SendMsgToChar("Уже открыто!\r\n", ch);
-		else if (!(DOOR_IS_LOCKED(ch, obj, door)) && IS_SET(flags_door[subcmd], kNeedLocked))
-			SendMsgToChar("Да отперли уже все...\r\n", ch);
-		else if (!(DOOR_IS_UNLOCKED(ch, obj, door)) && IS_SET(flags_door[subcmd], kNeedUnlocked))
-			SendMsgToChar("Угу, заперто.\r\n", ch);
-		else if (!HasKey(ch, keynum) && !privilege::CheckFlag(ch, privilege::kUseSkills)
-			&& ((subcmd == kScmdLock) || (subcmd == kScmdUnlock)))
-			SendMsgToChar("У вас нет ключа.\r\n", ch);
-		else if (DOOR_IS_BROKEN(ch, obj, door) && !privilege::CheckFlag(ch, privilege::kUseSkills)
-			&& ((subcmd == kScmdLock) || (subcmd == kScmdUnlock)))
-			SendMsgToChar("Замок сломан.\r\n", ch);
-		else if (ok_pick(ch, keynum, obj, door, subcmd))
-			do_doorcmd(ch, obj, door, (EDoorScmd) subcmd);
-	}
-}
-
-void do_stand(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/) {
-	if (ch->GetPosition() > EPosition::kSleep && AFF_FLAGGED(ch, EAffect::kSleep)) {
-		SendMsgToChar("Вы сладко зевнули и решили еще немного подремать.\r\n", ch);
-		act("$n сладко зевнул$a и решил$a еще немного подремать.", true, ch, nullptr, nullptr, kToRoom | kToArenaListen);
-		ch->SetPosition(EPosition::kSleep);
-	}
-
-	if (ch->IsOnHorse()) {
-		act("Прежде всего, вам стоит слезть с $N1.", false, ch, nullptr, ch->get_horse(), kToChar);
-		return;
-	}
-	switch (ch->GetPosition()) {
-		case EPosition::kStand: SendMsgToChar("А вы уже стоите.\r\n", ch);
-			break;
-		case EPosition::kSit: SendMsgToChar("Вы встали.\r\n", ch);
-			act("$n поднял$u.", true, ch, nullptr, nullptr, kToRoom | kToArenaListen);
-			// Will be sitting after a successful bash and may still be fighting.
-			ch->GetEnemy() ? ch->SetPosition(EPosition::kFight) : ch->SetPosition(EPosition::kStand);
-			break;
-		case EPosition::kRest: SendMsgToChar("Вы прекратили отдыхать и встали.\r\n", ch);
-			act("$n прекратил$g отдых и поднял$u.", true, ch, nullptr, nullptr, kToRoom | kToArenaListen);
-			ch->GetEnemy() ? ch->SetPosition(EPosition::kFight) : ch->SetPosition(EPosition::kStand);
-			break;
-		case EPosition::kSleep: SendMsgToChar("Пожалуй, сначала стоит проснуться!\r\n", ch);
-			break;
-		case EPosition::kFight: SendMsgToChar("Вы дрались лежа? Это что-то новенькое.\r\n", ch);
-			break;
-		default: SendMsgToChar("Вы прекратили летать и опустились на грешную землю.\r\n", ch);
-			act("$n опустил$u на землю.", true, ch, nullptr, nullptr, kToRoom | kToArenaListen);
-			ch->SetPosition(EPosition::kStand);
-			break;
-	}
-}
-
-void do_sit(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/) {
-	if (ch->IsOnHorse()) {
-		act("Прежде всего, вам стоит слезть с $N1.", false, ch, nullptr, ch->get_horse(), kToChar);
-		return;
-	}
-	switch (ch->GetPosition()) {
-		case EPosition::kStand: SendMsgToChar("Вы сели.\r\n", ch);
-			act("$n сел$g.", false, ch, nullptr, nullptr, kToRoom | kToArenaListen);
-			ch->SetPosition(EPosition::kSit);
-			break;
-		case EPosition::kSit: SendMsgToChar("А вы и так сидите.\r\n", ch);
-			break;
-		case EPosition::kRest: SendMsgToChar("Вы прекратили отдыхать и сели.\r\n", ch);
-			act("$n прервал$g отдых и сел$g.", true, ch, nullptr, nullptr, kToRoom | kToArenaListen);
-			ch->SetPosition(EPosition::kSit);
-			break;
-		case EPosition::kSleep: SendMsgToChar("Вам стоит проснуться.\r\n", ch);
-			break;
-		case EPosition::kFight: SendMsgToChar("Сесть? Во время боя? Вы явно не в себе.\r\n", ch);
-			break;
-		default: SendMsgToChar("Вы прекратили свой полет и сели.\r\n", ch);
-			act("$n прекратил$g свой полет и сел$g.", true, ch, nullptr, nullptr, kToRoom | kToArenaListen);
-			ch->SetPosition(EPosition::kSit);
-			break;
-	}
-}
-
-void do_rest(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/) {
-	if (ch->IsOnHorse()) {
-		act("Прежде всего, вам стоит слезть с $N1.", false, ch, nullptr, ch->get_horse(), kToChar);
-		return;
-	}
-	switch (ch->GetPosition()) {
-		case EPosition::kStand: SendMsgToChar("Вы присели отдохнуть.\r\n", ch);
-			act("$n присел$g отдохнуть.", true, ch, nullptr, nullptr, kToRoom | kToArenaListen);
-			ch->SetPosition(EPosition::kRest);
-			break;
-		case EPosition::kSit: SendMsgToChar("Вы пристроились поудобнее для отдыха.\r\n", ch);
-			act("$n пристроил$u поудобнее для отдыха.", true, ch, nullptr, nullptr, kToRoom | kToArenaListen);
-			ch->SetPosition(EPosition::kRest);
-			break;
-		case EPosition::kRest: SendMsgToChar("Вы и так отдыхаете.\r\n", ch);
-			break;
-		case EPosition::kSleep: SendMsgToChar("Вам лучше сначала проснуться.\r\n", ch);
-			break;
-		case EPosition::kFight: SendMsgToChar("Отдыха в бою вам не будет!\r\n", ch);
-			break;
-		default: SendMsgToChar("Вы прекратили полет и присели отдохнуть.\r\n", ch);
-			act("$n прекратил$g полет и пристроил$u поудобнее для отдыха.", false, ch, nullptr, nullptr, kToRoom | kToArenaListen);
-			ch->SetPosition(EPosition::kRest);
-			break;
 	}
 }
